void

+++ SUBS/savefig.m
@@ -4,38 +4,28 @@
 % Matlab:  7.9
 % Author:  NK
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-function savefig(outdir,rez,xdim,ydim,tit,zbub)
-%% set up figure
-if nargin<6
-    zbub=false;
-end
-set(gcf,'renderer','opengl');
-resolution=get(0,'ScreenPixelsPerInch');
-xdim=xdim*rez/resolution;
-ydim=ydim*rez/resolution;
-set(gcf,'paperunits','inch','papersize',[xdim ydim]/rez,'paperposition',[0 0 [xdim ydim]/rez]);
-mkdirp(outdir);
-fname=[outdir,tit];
-
-% fnamepdf=[outdir,tit,'.pdf'];
-
-% if zbub    
-%     set(gcf,'renderer','zbuffer');
-disp(['printing' fname ' to ps']);
-    print(gcf, '-dpsc2', [fname,'.ps'])
-     
-eval(['print ',[fname,'.eps'] , ' -f -r',num2str(rez),' -depsc ;'])
- system(['epstopdf ' fname '.eps']);
- system(['pdfcrop ' fname '.pdf ' fname 'Crpd.pdf']);
- system(['rm ' fname '.eps']);
- system(['rm ' fname '.pdf']);
-    
-    
-%     system(['convert -quality 100 -density 30 ' fnamepng,' ',fnamepdf]);
-% else
-%    
-%    
-% end
-close(gcf);
+function savefig(outdir,rez,xdim,ydim,tit)
+	fname=[outdir,tit];
+	mkdirp(outdir);	
+	if rez==42
+		%% quick hack
+		disp(['quick print to ' [fname,'.png']])
+		print(gcf, '-dpng', [fname,'.png'])
+	else
+		%% set up figure
+		set(gcf,'renderer','opengl');
+		resolution=get(0,'ScreenPixelsPerInch');
+		xdim=xdim*rez/resolution;
+		ydim=ydim*rez/resolution;
+		set(gcf,'paperunits','inch','papersize',[xdim ydim]/rez,'paperposition',[0 0 [xdim ydim]/rez]);
+		disp(['printing' fname ' to ps']);
+		print(gcf, '-dpsc2', [fname,'.ps'])
+		eval(['print ',[fname,'.eps'] , ' -f -r',num2str(rez),' -depsc ;'])
+		system(['epstopdf ' fname '.eps']);
+		system(['pdfcrop ' fname '.pdf ' fname 'Crpd.pdf']);
+		system(['rm ' fname '.eps']);
+		system(['rm ' fname '.pdf']);
+	end
+	close(gcf);
 end
